+++ lib/hephaestus_oban/migrations/postgres/v02.ex
@@ -3,14 +3,14 @@ defmodule HephaestusOban.Migrations.Postgres.V02 do
 
   use Ecto.Migration
 
-  def up(_opts) do
-    alter table(:hephaestus_step_results) do
+  def up(%{prefix: prefix}) do
+    alter table(:hephaestus_step_results, prefix: prefix) do
       add_if_not_exists(:metadata_updates, :map, null: false, default: %{})
     end
   end
 
-  def down(_opts) do
-    alter table(:hephaestus_step_results) do
+  def down(%{prefix: prefix}) do
+    alter table(:hephaestus_step_results, prefix: prefix) do
       remove_if_exists(:metadata_updates, :map)
     end
   end