+++ lib/hephaestus_oban/migrations/postgres/v03.ex
@@ -3,14 +3,14 @@ defmodule HephaestusOban.Migrations.Postgres.V03 do
 
   use Ecto.Migration
 
-  def up(_opts) do
-    alter table(:hephaestus_step_results) do
+  def up(%{prefix: prefix}) do
+    alter table(:hephaestus_step_results, prefix: prefix) do
       add_if_not_exists(:workflow_version, :integer, null: false, default: 1)
     end
   end
 
-  def down(_opts) do
-    alter table(:hephaestus_step_results) do
+  def down(%{prefix: prefix}) do
+    alter table(:hephaestus_step_results, prefix: prefix) do
       remove_if_exists(:workflow_version, :integer)
     end
   end